Moved club creation logic to new service file
