+++ app/Http/Controllers/ClubCreationController.php
@@ -3,26 +3,25 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use App\Services\ClubService;
+use App\Services\ClubCreationService;
 
 class ClubCreationController extends Controller
 {
-    protected $clubService;
-    protected $notificationService;
+    protected $clubCreationService;
 
-    public function __construct(ClubService $clubService) {
-        $this->clubService = $clubService;
+    public function __construct(ClubCreationService $clubCreationService) {
+        $this->clubCreationService = $clubCreationService;
     }
 
     public function addNewClub(Request $request) {
-        return $this->clubService->handleAddNewClub($request);
+        return $this->clubCreationService->handleAddNewClub($request);
     }
 
     public function makeNewClubCreationRequest(Request $request) {
-        return $this->clubService->handleNewClubCreationRequest($request);
+        return $this->clubCreationService->handleNewClubCreationRequest($request);
     }
 
     public function fetchRequestsPage(Request $request) {
-        return $this->clubService->prepareAndRenderRequestsView($request);
+        return $this->clubCreationService->prepareAndRenderRequestsView($request);
     }
 }
